pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spongebobs-corpse"
version = "0.1.0"
description = "Reproducible impact attenuation analysis pipeline for viscoelastic pad experiments"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "cytronicoder" }]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.1",
  "matplotlib>=3.8",
  "scipy>=1.11",
  "pyyaml>=6.0",
]

[project.scripts]
spongebobs-corpse = "spongebobs_corpse.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "mypy>=1.10",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
